refactor: refactor model hook decorators
